- [x] CHANGELOG.md updated with a link to the PR (not the Issue)
  - [x] Rebased/mergable
  - [x] Tests pass

Connect #1687 

### The problem

The UI would show, at most, 100 alert rules even if there were more

### The Solution

This introduces a wrapper around the Kapacitor client which drives the `limit` and `offset` parameters of the Kapacitor API automatically when it receives a `limit` and `offset` of 0 for its `ListTasks` method. The result is that all rules are returned, even when multiple requests are necessary to fetch them all.

Additional test coverage has been added around the `KapacitorRulesGet` action as well, where there was none previously.