@goller 2017-08-17
I would much rather see links with `next` and `prev` exposed in the API rather than `limit` and `offset` with special behavior.

The next/prev links would encode the limit/offset into kapacitor.  To find out how may tasks there are in kapacitor at high speed, run this until you get zero tasks returned: 

```curl "localhost:9092/kapacitor/v1/tasks?fields=id&offset=100"```

If there is no more tasks then no next link.