@timraymond 2017-08-22
Before concurrently fetching Kapacitor results:

```
BenchmarkKapaClient100-8           50000             25095 ns/op
BenchmarkKapaClient1000-8           1000           1515398 ns/op
BenchmarkKapaClient10000-8            10         137584474 ns/op
BenchmarkKapaClient100000-8            1        13188801433 ns/op
```

After:

```
BenchmarkKapaClient100-8           50000             40137 ns/op
BenchmarkKapaClient1000-8           1000           1655136 ns/op
BenchmarkKapaClient10000-8            20          89788016 ns/op
BenchmarkKapaClient100000-8            1        9328280331 ns/op
```

Throughput is slightly better with >10,000 tasks, with lower latency. However there is a cost of slightly higher latency with fewer tasks. This cost is measured in microseconds where the savings with larger workloads is measured in milliseconds, so I call this one a win.

With an accurate (and not network-intensive) way of counting the total number of tasks available, we could pre-allocate the output buffer of tasks and reduce the allocations required to `append` to it and bring these numbers down further. Most of the remaining time spent is related to GC and copying buffers around related to that.