@timraymond 2017-08-22
This should be ready for another look. /cc @jaredscheib 

@goller I agree that your proposal has merits, in that we can incrementally hydrate the list of tasks with multiple requests in the frontend before users can scroll/paginate to see them. Meanwhile the count we separately provide would be correct since we could figure that out in advance as you mention. However, I believe that this should be pursued as another issue and advocated for in IPM. This PR  requires no modifications to the frontend to fix #1687 where your proposal would, which will delay this fix further.