@jaredscheib 2017-08-23
Seems to work well. Good parallelization using a worker pool via goroutines and channels.

What do you think about factoring out the goroutines into named exported functions and testing them? Or otherwise somehow testing the parallelization specifically? The client is being tested on the whole, but that stuff doesn't seem to be.

Added review comments above. (Somehow this one didn't make it through with those.)

I tested this on 151 alerts via the Chronograf client and it worked, delivering all 151 at once. Success! Though it was a bit slow at ~2s roundtrip. Any thoughts on how that could be made faster?

Functionally this looks like an LGTM to me, though I'd defer to @goller in case he has comments on the code as-is. In my comments above, I do include some questions about whether the code could possibly be a bit cleaner, and test coverage for the async stuff could be implemented. Those would be my main two asks prior to merge.