@jaredscheib 2017-08-23
In speaking with @timraymond, here's the final checklist before merging:
1. [ ] the taskAsm `1` val as constant
1. [ ] the var `taskAsm`
1. [ ] break up the long function
1. [ ] delete the third test helper case, since it's not being used